feat(frontend): refactor mood components and enhance past sprints page
- Extract MoodTrendChart and MoodGridDisplay into pure presentational components
- Implement MoodTrendContainer and MoodGridContainer (via existing widgets)
- Add PastSprintDetails component for lazy loading of historical data
- Redesign PastSprintsPage with MUI Accordions and multi-criteria sorting
- Add i18n translations for "Mood Grid" labels
