+++ app/frontend/src/i18n/locales/fr.json
@@ -97,7 +97,8 @@
     "failedLoad": "Échec du chargement des données. Assurez-vous d'être connecté.",
     "noSprintsFound": "Aucun sprint passé trouvé.",
     "team": "Équipe : {{name}}",
-    "from": "Du {{start}} au {{end}}"
+    "from": "Du {{start}} au {{end}}",
+    "moodGrid": "Grille d'humeurs"
   },
   "adminUsers": {
     "title": "Administration des Utilisateurs",
